requirements.txt
aiohttp
requests
pyotp
apscheduler
aiofiles
pystray
Pillow